strawberry-backend/requirements.txt
django==4.2.6
psycopg2===2.9.9
django-graphql-jwt==0.4.0
requests==2.31.0
django-cors-headers==4.3.0
gunicorn==20.0.4
tzdata==2022.7
whitenoise==6.6.0
strawberry-graphql-django==0.21.0
strawberry-django-optimizer==0.1.0